+++ apps/desktop-wallet/src/api/apiDataHooks/address/useFetchAddressTokensByType.ts
@@ -16,7 +16,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { UseFetchAddressProps } from '@/api/apiDataHooks/address/types'
+import { UseFetchAddressProps } from '@/api/apiDataHooks/address/addressApiDataHooksTypes'
 import useFetchAddressBalances from '@/api/apiDataHooks/address/useFetchAddressBalances'
 import useFetchSeparatedTokensByType from '@/api/apiDataHooks/utils/useFetchSeparatedTokensByType'
