+++ apps/desktop-wallet/src/api/apiUtils.ts
